% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abstract_tokens}
\alias{abstract_tokens}
\title{Pre-computed abstract tokens}
\format{A data frame with two variables: \code{application.id}, \code{word}.}
\source{
Computed from the ABSTRACTS tables. Abstracts are tokenized and stop
words are removed.
}
\usage{
abstract_tokens
}
\description{
Pre-computed abstract tokens
}
\keyword{datasets}
