% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abstract_words}
\alias{abstract_words}
\title{Tokenized words from abstracts.}
\format{
A tibble with five variables: \code{activity}, \code{fiscal_year}, \code{institute}, \code{word}, \code{n}.
}
\source{
\url{https://reporter.nih.gov/exporter/abstracts}
}
\usage{
abstract_words
}
\description{
Tokenized words from abstracts.
}
\keyword{datasets}
