% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clinical_studies}
\alias{clinical_studies}
\title{Clinical trial information.}
\format{
A data frame with three variables: \code{project.num}, \code{trial.id} and \code{study.status}
}
\source{
\url{https://reporter.nih.gov/exporter/clinicalstudies}
}
\usage{
clinical_studies
}
\description{
Clinical trial information.
}
\keyword{datasets}
