% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_vars.R
\docType{data}
\name{nih.institutes}
\alias{nih.institutes}
\title{nih.institutes}
\format{An object of class \code{character} of length 27.}
\source{
\url{http://grants.nih.gov/grants/acronym_list.htm}
}
\usage{
nih.institutes
}
\description{
Charactor vector containing 27 NIH institutes in two-letter format
}
\keyword{datasets}

