% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{nihexporter-db}
\alias{copy_nihexporter}
\alias{has_nihexporter}
\alias{nihexporter-db}
\alias{nihexporter_postgres}
\alias{nihexporter_sqlite}
\title{Database versions of the nihexporter data}
\usage{
nihexporter_sqlite(path = NULL)

nihexporter_postgres(dbname = "nihexporter", ...)

has_nihexporter(type = c("sqlite", "postgresql"), ...)

copy_nihexporter(src, ...)
}
\arguments{
\item{path}{location of sqlite database file}

\item{dbname, ...}{Arguments passed on to \code{\link{src_postgres}}}
}
\description{
These functions cache the data from the \code{nihexporter} database in
a local database for use in examples and vignettes. Indexes are created
to make joining tables on natural keys efficient.
}
\keyword{internal}

