% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patents}
\alias{patents}
\title{Table linking patent IDs to project IDs.}
\format{
A data frame with two variables: \code{patent.id}, \code{project.num}
}
\source{
\url{http://exporter.nih.gov/ExPORTER_Catalog.aspx?sid=5&index=3}
}
\usage{
patents
}
\description{
Table linking patent IDs to project IDs.
}
\keyword{datasets}
