% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project.orgs}
\alias{project.orgs}
\title{project.orgs}
\format{A data frame with 4 variables: \code{org.duns},
  \code{org.city}, \code{org.state} and \code{org.name}

The org.duns is a unique identifier that links the `project.orgs` and
`projects` tables}
\usage{
project.orgs
}
\description{
project.orgs
}
\keyword{datasets}

