% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project_io}
\alias{project_io}
\title{Pre-computed total costs, number of patents and publications per project.}
\format{
A data frame with four variables: \code{project.num},
\code{n.pubs}, \code{n.patents} and \code{total.cost}
}
\source{
Computed from \link{projects} table.
}
\usage{
project_io
}
\description{
Pre-computed total costs, number of patents and publications per project.
}
\keyword{datasets}
