% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project_output}
\alias{project_output}
\title{project_output}
\format{A data frame with three variables: \code{project.num},
  \code{n.pubs}, \code{n.patents} and \code{project.cost}}
\usage{
project_output
}
\description{
Pre-computed project.cost, number of patents and publications associated with each
project
}
\keyword{datasets}

