% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project_pis}
\alias{project_pis}
\title{project_pis}
\format{A data frame with 2 variables: \code{project.num} and \code{pi.id}

The pi.id is a unique identifier that can be used to identify specific
PIs in NIH REPORTER.}
\source{
\url{http://exporter.nih.gov/ExPORTER_Catalog.aspx?sid=4&index=0}
}
\usage{
project_pis
}
\description{
project_pis
}
\keyword{datasets}

