% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projects}
\alias{projects}
\title{Project information.}
\format{
A data frame with 19 variables
}
\source{
\url{https://reporter.nih.gov/exporter/projects}
}
\usage{
projects
}
\description{
Project information.
}
\keyword{datasets}
