% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projects_min}
\alias{projects_min}
\title{A minimum set of project information from 2006-2024.}
\format{
A data frame with 14 variables
}
\source{
\url{https://reporter.nih.gov/exporter/projects}
}
\usage{
projects_min
}
\description{
A minimum set of project information from 2006-2024.
}
\keyword{datasets}
