% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{publications}
\alias{publications}
\title{publications}
\format{A data frame with two variables: \code{pmid}, \code{pmc.id},
 \code{journal.title.abbr}, \code{pub.year}}
\source{
\url{http://exporter.nih.gov/ExPORTER_Catalog.aspx?sid=0&index=4}
}
\usage{
publications
}
\description{
Information about specific publications
}
\keyword{datasets}

