% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{publinks}
\alias{publinks}
\title{Table linking PubMed IDs to project IDs.}
\format{
A data frame with two variables: \code{pmid}, \code{project.num}
}
\source{
\url{https://reporter.nih.gov/exporter/linktables}
}
\usage{
publinks
}
\description{
Table linking PubMed IDs to project IDs.
}
\keyword{datasets}
