% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper_vars.R
\docType{data}
\name{nih.institutes}
\alias{nih.institutes}
\title{nih.institutes}
\format{\preformatted{ chr [1:27] "AA" "AG" "AI" "AR" "AT" "CA" "DA" "DC" "DE" ...
}}
\source{
\url{http://grants.nih.gov/grants/acronym_list.htm}
}
\usage{
nih.institutes
}
\description{
Charactor vector containing 27 NIH institutes in two-letter format
}
\keyword{datasets}

