% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_vars.R
\docType{data}
\name{nih.institutes}
\alias{nih.institutes}
\title{Two-letter acronyms for NIH institutes}
\format{An object of class \code{character} of length 27.}
\source{
\url{http://grants.nih.gov/grants/acronym_list.htm}
}
\usage{
nih.institutes
}
\description{
Two-letter acronyms for NIH institutes
}
\keyword{datasets}

