% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patents}
\alias{patents}
\title{patents}
\format{A data frame with two variables: \code{patent.id}, \code{project.num}}
\source{
\url{http://exporter.nih.gov/ExPORTER_Catalog.aspx?sid=0&index=3}
}
\usage{
patents
}
\description{
Links patent IDs to Project IDs.
}
\keyword{datasets}

