% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project_io}
\alias{project_io}
\title{project_io}
\format{A data frame with four variables: \code{project.num},
  \code{n.pubs}, \code{n.patents} and \code{total.cost}}
\source{
pre-computed from \code{projects} table
}
\usage{
project_io
}
\description{
Pre-computed project.cost, number of patents and publications associated with each
project
}
\keyword{datasets}

