% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project_orgs}
\alias{project_orgs}
\title{Table linking organizations to applications.}
\format{A data frame with 2 variables: \code{org.duns} and
  \code{application.id} used to link specific applications to organizations}
\source{
\url{http://exporter.nih.gov/ExPORTER_Catalog.aspx?sid=3&index=0}
}
\usage{
project_orgs
}
\description{
Table linking organizations to applications.
}
\keyword{datasets}

