% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project_pis}
\alias{project_pis}
\title{Principal investigators.}
\format{
A data frame with 2 variables: \code{project.num} and \code{pi.id}

The \code{pi.id} is a unique identifier that can be used to identify specific
principal investigators in NIH REPORTER.
}
\source{
\url{https://reporter.nih.gov/exporter/projects}
}
\usage{
project_pis
}
\description{
Principal investigators.
}
\keyword{datasets}
