% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projects}
\alias{projects}
\title{projects}
\format{A data frame with 13 variables:
\describe{
\item{\code{project.num}}{core project number, e.g. R01GM1234556}
\item{\code{activity}}{activity code, e.g. R01}
\item{\code{application.type}}{one-digit code to identify type of application funded}
\item{\code{arra.funded}}{'Y' indicates ARRA funded}
\item{\code{suffix}}{'A'mended and 'S'upplmental + serial number}
\item{\code{institute}}{administering institute}
\item{\code{fiscal.year}}{year of funding}
\item{\code{org.duns}}{DUNS number}
\item{\code{project.start}}{in ymd format}
\item{\code{project.end}}{in ymd format}
\item{\code{study.section}}{3 letter code}
\item{\code{total.cost}}{total (direct + indirect) for fiscal year}
}}
\source{
\url{http://exporter.nih.gov/ExPORTER_Catalog.aspx?sid=4&index=0}
}
\usage{
projects
}
\description{
projects
}
\keyword{datasets}

