% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projects}
\alias{projects}
\title{Project information.}
\format{
A data frame with 19 variables
}
\source{
\url{http://exporter.nih.gov/ExPORTER_Catalog.aspx?sid=3&index=0}
}
\usage{
projects
}
\description{
Project information.
}
\keyword{datasets}
