% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projects}
\alias{projects}
\title{projects}
\format{A data frame with 11 variables: \code{application.id}, \code{institute},
  \code{activity}, \code{project.num}, \code{fiscal.year},
  \code{org.duns}, \code{suffix}, \code{project.start},
  \code{project.end}, \code{study.section}, \code{total.cost}}
\source{
\url{http://exporter.nih.gov/ExPORTER_Catalog.aspx?sid=4&index=0}
}
\usage{
projects
}
\description{
projects
}
\keyword{datasets}

