% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projects}
\alias{projects}
\title{NIH EXPORTER}
\format{A data frame with nine variables: \code{application.id}, \code{institute},
  \code{activity}, \code{project.num}, \code{fiscal.year}, \code{project.start},
  \code{project.end}, \code{study.section}, \code{total.cost}

For further details, see \url{http://exporter.nih.gov/ExPORTER_Catalog.aspx?sid=4&index=0}}
\usage{
projects
}
\description{
Minimal set of data provided by NIH EXPORTER describing funded biomedical research projects from 1985-2014.
}
\details{
projects
}
\keyword{datasets}

