% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{publications}
\alias{publications}
\title{Publication information.}
\format{A data frame with two variables: \code{pmid}, \code{pmc.id},
 \code{journal.title.abbr}, \code{pub.year}, \code{rcr}}
\source{
\url{https://exporter.nih.gov/ExPORTER_Catalog.aspx?sid=4&index=2}

\url{https://icite.od.nih.gov/}
}
\usage{
publications
}
\description{
Publication information.
}
\keyword{datasets}

