% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{publications}
\alias{publications}
\title{Publication information.}
\format{
A data frame with 4 variables: \code{pmid}, \code{pmc.id},
\code{pub.year}, \code{rcr}
}
\source{
\url{https://exporter.nih.gov/ExPORTER_Catalog.aspx?sid=4&index=2}

\url{https://icite.od.nih.gov/}
}
\usage{
publications
}
\description{
Publication information.
}
\note{
\code{rcr} values change over time. Values in this table were
downloaded on 31 Dec 2016. Use \code{\link[=rcr]{rcr()}} to obtain current
values for select \code{pmid} values.
}
\keyword{datasets}
