% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{publinks}
\alias{publinks}
\title{publinks}
\format{A data frame with two variables: \code{pmid}, \code{project.num}

For further details, see \url{http://exporter.nih.gov/ExPORTER_Catalog.aspx?sid=0&index=4}}
\usage{
publinks
}
\description{
Links Pubmed IDs to Project IDs.
}
\keyword{datasets}

